# Conflicts:
#	flickit-assessment-common/src/main/resources/i18n/kit/messages_en.properties
#	flickit-assessment-data/src/main/resources/liquibase/changelog-0.29.xml
#	flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/common/ErrorMessageKey.java
#	flickit-assessment-kit/src/test/resources/postman/flickit-assessment-kit.postman_collection.json

+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/out/persistence/maturitylevel/MaturityLevelPersistenceJpaAdapter.java
@@ -1,6 +1,7 @@
 package org.flickit.assessment.kit.adapter.out.persistence.maturitylevel;
 
 import lombok.RequiredArgsConstructor;
+import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.data.jpa.kit.levelcompetence.LevelCompetenceJpaEntity;
 import org.flickit.assessment.data.jpa.kit.levelcompetence.LevelCompetenceJpaRepository;
 import org.flickit.assessment.data.jpa.kit.maturitylevel.MaturityLevelJpaEntity;
@@ -8,6 +9,7 @@ import org.flickit.assessment.data.jpa.kit.maturitylevel.MaturityLevelJpaEntity.
 import org.flickit.assessment.data.jpa.kit.maturitylevel.MaturityLevelJpaRepository;
 import org.flickit.assessment.kit.application.domain.MaturityLevel;
 import org.flickit.assessment.kit.application.domain.MaturityLevelCompetence;
+import org.flickit.assessment.kit.application.domain.MaturityLevelOrder;
 import org.flickit.assessment.kit.application.port.out.maturitylevel.*;
 import org.springframework.stereotype.Component;
 
@@ -19,6 +21,7 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 
 import static org.flickit.assessment.kit.adapter.out.persistence.maturitylevel.MaturityLevelMapper.mapToJpaEntityToPersist;
+import static org.flickit.assessment.kit.common.ErrorMessageKey.MATURITY_LEVEL_ID_NOT_FOUND;
 
 @Component
 @RequiredArgsConstructor
@@ -39,11 +42,13 @@ public class MaturityLevelPersistenceJpaAdapter implements
 
     @Override
     public void delete(Long id, Long kitVersionId) {
+        if(!repository.existsByIdAndKitVersionId(id, kitVersionId))
+            throw new ResourceNotFoundException(MATURITY_LEVEL_ID_NOT_FOUND);
         repository.deleteByIdAndKitVersionId(id, kitVersionId);
     }
 
     @Override
-    public void update(List<MaturityLevel> maturityLevels, Long kitVersionId, UUID lastModifiedBy) {
+    public void updateAll(List<MaturityLevel> maturityLevels, Long kitVersionId, UUID lastModifiedBy) {
         Map<EntityId, MaturityLevel> idToModel = maturityLevels.stream()
             .collect(Collectors.toMap(
                 ml -> new EntityId(ml.getId(), kitVersionId),
@@ -61,6 +66,36 @@ public class MaturityLevelPersistenceJpaAdapter implements
         });
         repository.saveAll(entities);
         repository.flush();
+    }
+
+    @Override
+    public void update(MaturityLevel maturityLevel, Long kitVersionId, LocalDateTime lastModificationTime, UUID lastModifiedBy ) {
+        if (!repository.existsByIdAndKitVersionId(maturityLevel.getId(), kitVersionId))
+            throw new ResourceNotFoundException(MATURITY_LEVEL_ID_NOT_FOUND);
+
+        repository.update(maturityLevel.getId(), kitVersionId, maturityLevel.getTitle(), maturityLevel.getIndex(), maturityLevel.getCode(),
+            maturityLevel.getDescription(), maturityLevel.getValue(), lastModificationTime, lastModifiedBy);
+    }
+
+    @Override
+    public void updateOrders(List<MaturityLevelOrder> maturityLevelOrders, Long kitVersionId, UUID lastModifiedBy) {
+        Map<EntityId, MaturityLevelOrder> idToModel = maturityLevelOrders.stream()
+            .collect(Collectors.toMap(
+                ml -> new EntityId(ml.getId(), kitVersionId),
+                ml -> ml
+            ));
+        List<MaturityLevelJpaEntity> entities = repository.findAllById(idToModel.keySet());
+        if (entities.size() != maturityLevelOrders.size())
+            throw new ResourceNotFoundException(MATURITY_LEVEL_ID_NOT_FOUND);
+
+        entities.forEach(x -> {
+            MaturityLevelOrder newLevel = idToModel.get(new EntityId(x.getId(), kitVersionId));
+            x.setIndex(newLevel.getIndex());
+            x.setValue(newLevel.getValue());
+            x.setLastModificationTime(LocalDateTime.now());
+            x.setLastModifiedBy(lastModifiedBy);
+        });
+        repository.saveAll(entities);
     }
 
     @Override
